refactor(wizard): name the spec configurator for what it is
`spec_flow` sat one word away from `integrations/setup_flow` and said
nothing about which layer it was on, which is a confusing pair to leave
behind for the remaining 41 integrations.

`configurators/spec_configurator.py` matches how its siblings in that
package are named, and the docstring now states the split outright: this
module collects values from the user, setup_flow decides where they are
persisted.
