+++ surfaces/cli/wizard/configurators/chat_notifications.py
@@ -15,7 +15,7 @@ from surfaces.cli.wizard._ui import (
     _render_integration_result,
     _string_value,
 )
-from surfaces.cli.wizard.configurators.spec_flow import configure_from_spec
+from surfaces.cli.wizard.configurators.spec_configurator import configure_from_spec
 from surfaces.cli.wizard.integration_health import (
     validate_discord_bot,
     validate_rocketchat,
